+++ app/Http/Controllers/API/TagController.php
@@ -10,12 +10,17 @@ use App\Http\Resources\Tag as TagResource;
 
 class TagController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth:api')->except(['index', 'show']);
+    }
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        $tag = TagResource::collection(Tag::all());
+        $limit = $request->input('limit') <= 50 ? $request->input('limit') : 15;
+        $tag = TagResource::collection(Tag::paginate($limit));
         return $tag->response()->setStatusCode(200);
     }
 
@@ -43,6 +48,8 @@ class TagController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $idtag = Tag::findOrFail($id);
+        $this->authorize('update', $idtag);
         $tag = new TagResource(Tag::findOrFail($id));
         $tag->update($request->all());
 
@@ -54,6 +61,8 @@ class TagController extends Controller
      */
     public function destroy($id)
     {
+        $idtag = Tag::findOrFail($id);
+        $this->authorize('delete', $idtag);
         Tag::find($id)->delete();
 
         return 204;
